Variant filtering implemented, filling histograms & pie charts with data
The variant filtering required making changes to the Cineca variant
JSON, to make it searchable using a filter.

Data was added to the Gender pie chart and to the Sample type bar
chart.

Additionally, ethnicity data was added to Patients. This is not
available in regular BBMRI or DKTK FHIR profiles, but is present in
the Cineca test dataset provided with the Beacon RI.