Clustered all entry type code together
The long-term goal for this repository is to make it possible for people
to easily add new entry types or delete old ones.

Hence, I have created separate packages for each of the entry types
individuals, biosamples and genomicVariations.

Each of these contains all of the code needed to run a single
entry type.