Modified Dockerfile, changed default site
The Dockerfile has been changed so that it first builds the Java project
with Maven. This is to fit in with the Github pipelines that I copied
from TransFAIR.

I also switched the default Beacon site from local Beacon to EGA Cineca.
That means that the component is ready to go straight away, there
is no need to first set up a local Beacon if all you want to do is
try it out.