Fine-tuned README to fit with the lens-beacon documentation
The README was initially written with the assumption that this component
would be used directly together with the Lens toolkit.

However, the toolkit has been split off into its own repo and a separate
lens-beacon repo has been created specifically for the Beacon GUI.

The README has been rewritten to reflect this change.